% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_name_that_tune}
\alias{run_name_that_tune}
\title{Name That Tune Application}
\usage{
run_name_that_tune(client_id, client_secret, port = getOption("shiny.port"))
}
\arguments{
\item{client_id}{character string of length 32. Spotify `Client ID`. See
details for more information.}

\item{client_secret}{character string of length 32. Spotify `Client Secret`.
See details for more information.}

\item{port}{The TCP port. See \code{\link[shiny]{runApp}}.}
}
\description{
Opens the Name That Tune game.
}
\details{
To get Client ID and Secret you need to have the account
on \href{https://developer.spotify.com/}{Spotify for Developers}.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

